p. 78 add Report class
It's back to the Design Patterns in Ruby work.
Everything else ongoing is stuck in branches at
at the moment and not ready to push. Design Patterns
makes a nice interlude while the other work proceeds.

This commit has the new refactored Report class
implemented, but no tests wrapped around it yet.
That will be a good topic for the next commit; I'm
writing this to help remember what to do next, since
it might be tomorrow, next week or next month.