p. 105 pass block to abserver
This is pretty interesting, and I'm not sure
I've run across it in the wild in quite so
simple a form, but it looks like something
which could be pretty useful.

The notion is to pass a block in which does the
actual updating. This is easy to do with our
custom Subject module, not sure how easy it
would be with Ruby's Observable built-in.

This commit has the block passing commented out
for now. It's not clear how to write it such that
passing either a block or an object would work
identically, without getting really ugly.
Something to bang on tomorrow and the next day.