fixed bug in spec, using File.readlines incorrectly
File.readlines returns an array on a given separator,
assumed to \n, but I was passing in 'r' which was making
the split on the r character. In other words, the computer
was working perfectly, doing exactly what I asked of it.

Which wasn't what I wanted.

Next up is fixing the date and time stamping spec.