p. 199 concluded specs for testing time stamping
For a given instant in (frozen) time, Time.new
and DateTime.now produce different strings. The
time is the same, but comparing the strings is
hard.

I've found it convenient to use DateTime.new
and pass in the exact parameters I want to
specify the time.

This is worthy of more study.