p. 231 reimplementing to move toward factory
The code so far in the chapter doesn't kend itself
to more than banal testing, so I haven't bothered.
As the chapter progresses, most of the early code is
going to be replaced anyway. I don't have a good
picture on how to rewrite this chapter better, but
I do have a feeling there is room for improvement.