p. 243 testing for IOFactory
This wraps up the coding for chapter 13 for now.

The lacks of specs reflects the inspectable nature
of the code which was written, and possibly some lack
of interest on my part for this material. The examples
felt very contrived.

It might be interesting in the future to create some
relevant examples, but as the author mentions in the
closing notes, there are generally better ways to handle
this problem in Ruby