p. 261 wrapping builder chapter
There didn't turn out to be a whole lot in this
chapter that I found super interesting. It was
a good review of techniques, but I suspect some
of these techniques would be more at home with c++
or java programmers. That's not a bad thing, and
it indicates opportunity for a fresh Ruby-oriented
overhaul.