ch15 code complete, lacks full specs
All the code for chapter 15 is punched in, but there
are about a dozen specs I want to fill out. This is
part of a new approach, where I punch in the boilerplate
first when I have larger blocks of time, then go through
and fill out specs when time is limited, for example,
before leaving for work.