p. 287, started packrat dsl
This chapter makes a jump in complexity by using the
file finding code written in the last chapter. This means
I need to rewrite all that file finding code into proper
Ruby structure; it won't work left in a giant rspec
autorun file. This will put me behind my anticipated
finish time for the book, but that's ok, as the material
is starting to get interesting.