p. 288 - rebuilding file finder
The backup dsl depends on the file finder working
correctly. This commit is a step towards rebuilding
all the material from the previous chapter to use
for the backup dsl.

I can already say I don't like using eval just for
the reason that having a bug in a script which is
called via eval gives no clue as to the point of error
in the script. Regardless of the security issues, this
is bad enough.