p. 289 backup script runs
For a very simple copy directive, the script runs.
The other directives for the backup script were
not evaluating correctly, so the new single.pr script
was added.

Note that some of the problems getting this to run
correctly are because of typos in the original material
from the previous chapter. This is a good case for
having done the requisite testing in that chapter
instead of punting it to later.

It would be smart to add testing to this chapter as
I go.