p. 289 switching gears to rebuilding rails
I'm almost finished with Design Patterns in Ruby,
by which I mean I have roughly 2.5 chapters remaining
to work through. At my current rate, that will take
a couple of months.

However, I've started into Rebuilding Rails, and I'm
going to purchase a Rails-specific programming course
online, so it makes sense to switch the dailies over
to Rails. It's what I'm getting paid for, and I want
to get much better at it.

I've put this book down in the past, and picked it up
again for the lastest effort, so I'm confident I'll
pick it back up and finish it some time in the future.