bug in the backup code
I'm now working my way through the material in the
the DSL chapter in some detail. I had been hoping
I'd be almost finished with this material by now, but
work responsibilities have resulted in very long days.
Hence back to rather smallish commits.

This material however is important enough that I want
to really understand what's it doing and how to do it,
which will take a fair bit of time.

Currently, there is some sort of bug where the files
to be backed up are not be read correctly. Running
`./packrat.rb` will stop the process with a pry about
where the problem first appears. Over the next couple
of days, will need to dig down from that point to
figure it out.