backup script now works
Fairly simple fix to the script, but I still need
to work through the entire implementation in more
detail.

What's problematic about this example is the entire
chapter hinges on calling `eval`. This is unacceptable
in production code, as the syntax is valid Ruby. I can
imagine someone sending an embedded `eval` through
the script, which could be bad.

A great challenge for going forward from the this chapter
is figuring out how to interpret the script without
running `eval`.