p. 298 starting ch 17 metaprogramming
The very last part of the dsl chapter was skipped
because it rewrote the application architecture.
I want both examples, but I don't have the cycles
at 4:18 am to figure out to engineer the repo for
it. I may return to it later.

The metaprogramming starts off with a simple
decorator.