That's it, that's all I care to do at this time.
For the current Chapter 18, I could use to add a
few more specs.

At some point in the future, it might be worth going
back through the book in more detail to both refresh
my knowledge, and see if I'm able to extract the
patterns in this book from whatever code I'm currently
working on. Or be able to apply any of the patterns to
code I'm currently working on.